restructure example
