+++ examples/model-serve-flow/01_Model_Compression/pyproject.toml
@@ -1,0 +1,11 @@
+[project]
+name = "Model_Compression"
+version = "0.1.0"
+description = "Compress a LLM using llmcompressor"
+requires-python = ">=3.12"
+
+dependencies = [
+    "torch==2.8.0",
+    "transformers==4.55.2",
+    "llmcompressor>=0.5.1",
+]
