+++ examples/model-serve-flow/02_Accuracy_Benchmarking/pyproject.toml
@@ -1,0 +1,10 @@
+[project]
+name = "Accuracy_Benchmarking"
+version = "0.1.0"
+description = "Evaluate the accuracy of a LLM (full + quantized)"
+requires-python = ">=3.12"
+
+dependencies = [
+    "torch==2.8.0",
+    "lm-eval[ifeval]>=0.4.8",
+]
