+++ examples/model-serve-flow/02_Accuracy_Benchmarking/utils.py
@@ -1,0 +1,90 @@
+import os
+import pickle
+from typing import Union
+
+import torch
+from lm_eval.evaluator import simple_evaluate
+
+
+def evaluate(
+    model_path: str,
+    tasks: list[str],
+    model: str = "hf",
+    num_fewshot: int = 0,
+    device: str = None,
+    limit: Union[int, float, None] = None,
+    apply_chat_template: bool = False,
+    verbosity: str = None,
+    log_samples: bool = False,
+    batch_size: Union[int, str] = None,
+):
+    """
+    Evaluate a language model on specified tasks using lm-eval-harness.
+
+    Parameters
+    ----------
+    model_path : str
+        Path to the pretrained model or HuggingFace model.
+    tasks : list[str]
+        List of task names to evaluate.
+    model : str, default "hf"
+        Selects which model type or provider is evaluated. default is "hf".
+    num_fewshot : int, default 0
+        Number of few-shot examples per task (0 for zero-shot).
+    device : str or None, default None
+        Device to run evaluation on ("cpu" or "cuda"). Auto-detected if None.
+    limit : int, float, or None, default None
+        Limit number of documents per task (int for count, float for fraction).
+    apply_chat_template : bool, default False
+        When True, adjusts delimiter handling for chat-style prompts:
+        sets the target delimiter to an empty string instead of the default whitespace.
+        Useful for likelihood or multiple-choice tasks with chat models to prevent
+        spacing issues between context and target. Does NOT apply a full chat template.
+    verbosity : str, default "DEBUG"
+        Level of logging for inputs and outputs. Can be set to None
+    log_samples : bool, default False
+        Whether to save sample outputs for debugging.
+
+    Returns
+    -------
+    dict
+        Evaluation results from simple_evaluate.
+    """
+
+    if device is None:  # check device
+        device = "cuda" if torch.cuda.is_available() else "cpu"
+    if limit is None:
+        limit = None if device == "cuda" else 4
+    if batch_size is None:
+        batch_size = "auto" if device == "cuda" else 4
+
+    model_args = {"pretrained": model_path}
+
+    try:
+        results = simple_evaluate(
+            model=model,
+            model_args=model_args,
+            tasks=tasks,
+            num_fewshot=num_fewshot,
+            limit=limit,
+            device=device,
+            apply_chat_template=apply_chat_template,
+            verbosity=verbosity,
+            log_samples=log_samples,
+            batch_size=batch_size,
+        )
+    except Exception as e:
+        raise RuntimeError("Could not run accuracy evaluation") from e
+
+    return results
+
+
+def save_pickle(path, data):
+    os.makedirs(path, exist_ok=True)
+    with open(f"{path}/results.pkl", "wb") as f:
+        pickle.dump(data, f)
+
+
+def load_pickle(path):
+    with open(f"{path}/results.pkl", "rb") as f:
+        return pickle.load(f)
