--- examples/model_serve_flow/requirements.txt
@@ -1,9 +1,0 @@
-torch
-transformers==4.55.2
-llmcompressor>=0.5.1
-protobuf
-sentencepiece
-vllm>=0.11.0
-lm-eval[ifeval]>=0.4.8
-openai==2.8.1
-guidellm
